requirements.txt
streamlit==1.32.0
pandas==2.2.1
deepseek-ai==0.0.1
python-dotenv==1.0.1
flask==3.0.2
flask-cors==4.0.0 